third_party/WebKit/LayoutTests/fast/css/margin-start-end-expected.txt

Tests margin-start and margin-end

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("ltr", "margin-inline-start", "10px", "width") is "90px"
PASS test("ltr", "margin-inline-end", "20px", "width") is "80px"
PASS test("ltr", "margin-inline-start", "10px", "margin-left") is "10px"
PASS test("ltr", "margin-inline-end", "20px", "margin-right") is "20px"
PASS test("ltr", "margin-left", "10px", "margin-inline-start") is "10px"
PASS test("ltr", "margin-right", "20px", "margin-inline-end") is "20px"
PASS test("rtl", "margin-inline-start", "10px", "width") is "90px"
PASS test("rtl", "margin-inline-end", "20px", "width") is "80px"
PASS test("rtl", "margin-inline-start", "10px", "margin-right") is "10px"
PASS test("rtl", "margin-inline-end", "20px", "margin-left") is "20px"
PASS test("rtl", "margin-right", "10px", "margin-inline-start") is "10px"
PASS test("rtl", "margin-left", "20px", "margin-inline-end") is "20px"
PASS successfullyParsed is true

TEST COMPLETE
